separate modules into different projects. Also added a shared folder with common code and dealt with setting up all the dependencies.
