--- Delivery 1 - Base Version/dida-gstore.csproj
@@ -1,9 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <RootNamespace>DIDA_GSTORE</RootNamespace>
-  </PropertyGroup>
-
-</Project>
