+++ Delivery 1 - Base Version/dida-gstore.sln
@@ -1,17 +1,44 @@
-
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DIDA-GSTORE", "dida-gstore.csproj", "{62D267E2-E993-43A2-934E-71DCBCE7D510}"
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30517.126
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PuppetMaster", "PuppetMaster\PuppetMaster.csproj", "{64619C7E-C3C8-4843-8633-3F24A470CEC0}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Server", "Server\Server.csproj", "{35C7285B-A804-4E01-8AA2-18DD3CC7C118}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Client", "Client\Client.csproj", "{FDBE174A-2424-4D35-B6DA-AB3355FA1852}"
+EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Shared", "Shared\Shared.shproj", "{FF7FB412-9FAB-4033-8E02-533148FEFE61}"
 EndProject
 Global
+	GlobalSection(SharedMSBuildProjectFiles) = preSolution
+		Shared\Shared.projitems*{35c7285b-a804-4e01-8aa2-18dd3cc7c118}*SharedItemsImports = 5
+		Shared\Shared.projitems*{64619c7e-c3c8-4843-8633-3f24a470cec0}*SharedItemsImports = 5
+		Shared\Shared.projitems*{fdbe174a-2424-4d35-b6da-ab3355fa1852}*SharedItemsImports = 5
+		Shared\Shared.projitems*{ff7fb412-9fab-4033-8e02-533148fefe61}*SharedItemsImports = 13
+	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{62D267E2-E993-43A2-934E-71DCBCE7D510}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{62D267E2-E993-43A2-934E-71DCBCE7D510}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{62D267E2-E993-43A2-934E-71DCBCE7D510}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{62D267E2-E993-43A2-934E-71DCBCE7D510}.Release|Any CPU.Build.0 = Release|Any CPU
+		{64619C7E-C3C8-4843-8633-3F24A470CEC0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{64619C7E-C3C8-4843-8633-3F24A470CEC0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{64619C7E-C3C8-4843-8633-3F24A470CEC0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{64619C7E-C3C8-4843-8633-3F24A470CEC0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{35C7285B-A804-4E01-8AA2-18DD3CC7C118}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{35C7285B-A804-4E01-8AA2-18DD3CC7C118}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{35C7285B-A804-4E01-8AA2-18DD3CC7C118}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{35C7285B-A804-4E01-8AA2-18DD3CC7C118}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FDBE174A-2424-4D35-B6DA-AB3355FA1852}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FDBE174A-2424-4D35-B6DA-AB3355FA1852}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FDBE174A-2424-4D35-B6DA-AB3355FA1852}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FDBE174A-2424-4D35-B6DA-AB3355FA1852}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {299D0C21-69D1-4B15-9455-030DD3F1DAC7}
 	EndGlobalSection
 EndGlobal
