--- Delivery 1 - Base Version/processCreationService/IProcessCreationService.cs
@@ -1,9 +1,0 @@
-namespace processCreationService
-{
-    public interface IProcessCreationService
-    {
-        void StartServer(string args);
-        void StartClient(string args);
-        void ShutdownAllProcesses();
-    }
-}
